refactored: utils `export, validate, parser json collection`
