+++ src/utils/json-parser.js
@@ -1,0 +1,16 @@
+function jsonParser(file) {
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.readAsText(file);
+    reader.onload = function () {
+      try {
+        const data = JSON.parse(reader.result);
+        resolve(data);
+      } catch (error) {
+        reject(error);
+      }
+    };
+  });
+}
+
+export default jsonParser;
